latest in projective dimension estimates, with per class accuracy

+++ feature_gen/sample.py
@@ -263,7 +263,8 @@ def save_features(outputs, path):
 def process_outputs(outputs):
     dataDict = {}
     for key in outputs:
-        data = np.array(outputs[key]).astype('float16')
+        data = np.array(outputs[key]).astype('float32')
+        #data = np.array(outputs[key]).astype('float16')
         og_shape = data.shape
         data = np.reshape(data, (data.shape[0]*data.shape[1], -1))
         key_ = key+'-ogshape_'+str(og_shape).replace(' ', '')
